Reduced indexId and generation in IndexController::RecordHandle struct from 64-bit integer down to 32-bit integer because realistically an index is not going to refresh more than 4 billion times while the app is open. This saves 8 * 2 = 16 bytes of memory per each search result, as there are currently 2 copies of search results - the search results then another copy as sorted search results

+++ kerything/IndexController.h
@@ -757,8 +757,8 @@ public:
     };
 
     struct RecordHandle {
-        quint64 indexId;
-        quint64 generation;
+        uint32_t indexId;
+        uint32_t generation;
         uint32_t recordIdx;
 
         // Index into DeviceIndex::mountPoints.
